this isn't really tested since I don't know how the snap store handles api keys, and I didn't want to get into a situation where it looked like I was taking credit as the author of your program.
I've been experiencing a problem on both ubuntu and suse where tgpt
seems to stop working every 24 hrs, and snaps can auto-update in the
background, whereas installing with go doesn't.

In order for this to work, you'll need to get a api key from the snap store and store it as a github secret.